Ejercicio 2, añadido la documentacion. Primer commit
